refactor: Remove legacy incidents wrapper
