+++ operate/client/src/App/ProcessInstance/IncidentsWrapper/index.tsx
@@ -13,56 +13,60 @@ import {incidentsStore} from 'modules/stores/incidents';
 import {observer} from 'mobx-react';
 import {Transition} from './styled';
 import {IncidentsFilter} from './IncidentsFilter';
-import {IncidentsTable as IncidentsTableV2} from './IncidentsTable/v2';
+import {IncidentsTable} from './IncidentsTable';
 import {PanelHeader} from 'modules/components/PanelHeader';
-import {getFilteredIncidents} from 'modules/utils/incidents';
+import {getFilteredIncidents, init} from 'modules/utils/incidents';
 import {useIncidents} from 'modules/hooks/incidents';
+import {type ProcessInstance} from '@vzeta/camunda-api-zod-schemas/8.8';
 
 type Props = {
+  processInstance: ProcessInstance;
   setIsInTransition: (isTransitionActive: boolean) => void;
 };
 
-const IncidentsWrapper: React.FC<Props> = observer(({setIsInTransition}) => {
-  const incidents = useIncidents();
-  const filteredIncidents = getFilteredIncidents(incidents);
+const IncidentsWrapper: React.FC<Props> = observer(
+  ({setIsInTransition, processInstance}) => {
+    const incidents = useIncidents();
+    const filteredIncidents = getFilteredIncidents(incidents);
 
-  useEffect(() => {
-    incidentsStore.init();
+    useEffect(() => {
+      init(processInstance);
 
-    return () => {
-      incidentsStore.reset();
-    };
-  }, []);
+      return () => {
+        incidentsStore.reset();
+      };
+    }, [processInstance]);
 
-  if (incidentsStore.incidentsCount === 0) {
-    return null;
-  }
+    if (incidentsStore.incidentsCount === 0) {
+      return null;
+    }
 
-  return (
-    <>
-      <Transition
-        in={incidentsStore.state.isIncidentBarOpen}
-        onEnter={() => setIsInTransition(true)}
-        onEntered={() => setIsInTransition(false)}
-        onExit={() => setIsInTransition(true)}
-        onExited={() => setIsInTransition(false)}
-        mountOnEnter
-        unmountOnExit
-        timeout={400}
-      >
-        <IncidentsOverlay>
-          <PanelHeader
-            title="Incidents View"
-            count={filteredIncidents.length}
-            size="sm"
-          >
-            <IncidentsFilter />
-          </PanelHeader>
-          <IncidentsTableV2 />
-        </IncidentsOverlay>
-      </Transition>
-    </>
-  );
-});
+    return (
+      <>
+        <Transition
+          in={incidentsStore.state.isIncidentBarOpen}
+          onEnter={() => setIsInTransition(true)}
+          onEntered={() => setIsInTransition(false)}
+          onExit={() => setIsInTransition(true)}
+          onExited={() => setIsInTransition(false)}
+          mountOnEnter
+          unmountOnExit
+          timeout={400}
+        >
+          <IncidentsOverlay>
+            <PanelHeader
+              title="Incidents View"
+              count={filteredIncidents.length}
+              size="sm"
+            >
+              <IncidentsFilter />
+            </PanelHeader>
+            <IncidentsTable />
+          </IncidentsOverlay>
+        </Transition>
+      </>
+    );
+  },
+);
 
 export {IncidentsWrapper};
